+++ data/book_map.csv
@@ -1,0 +1,67 @@
+﻿korean_name,book_id,full_name
+창,1,창세기
+출,2,출애굽기
+레,3,레위기
+민,4,민수기
+신,5,신명기
+수,6,여호수아
+삿,7,사사기
+룻,8,룻기
+삼상,9,사무엘상
+삼하,10,사무엘하
+왕상,11,열왕기상
+왕하,12,열왕기하
+대상,13,역대상
+대하,14,역대하
+스,15,에스라
+느,16,느헤미야
+에,17,에스더
+욥,18,욥기
+시,19,시편
+잠,20,잠언
+전,21,전도서
+아,22,아가
+사,23,이사야
+렘,24,예레미야
+애,25,예레미야 애가
+겔,26,에스겔
+단,27,다니엘
+호,28,호세아
+욜,29,요엘
+암,30,아모스
+옵,31,오바댜
+욘,32,요나
+미,33,미가
+나,34,나훔
+합,35,하박국
+습,36,스바냐
+학,37,학개
+슥,38,스가랴
+말,39,말라기
+마,40,마태복음
+막,41,마가복음
+눅,42,누가복음
+요,43,요한복음
+행,44,사도행전
+롬,45,로마서
+고전,46,고린도전서
+고후,47,고린도후서
+갈,48,갈라디아서
+엡,49,에베소서
+빌,50,빌립보서
+골,51,골로새서
+살전,52,데살로니가전서
+살후,53,데살로니가후서
+딤전,54,디모데전서
+딤후,55,디모데후서
+딛,56,디도서
+몬,57,빌레몬서
+히,58,히브리서
+약,59,야고보서
+벧전,60,베드로전서
+벧후,61,베드로후서
+요일,62,요한일서
+요이,63,요한이서
+요삼,64,요한삼서
+유,65,유다서
+계,66,요한계시록
