+++ data/qt_data.txt
@@ -1,0 +1,8 @@
+[12월 QT/통독 범위]            
+12.9 (월) 벧전2 / 골1-4
+12.10 (화) 벧전3 / 살전1-5
+12.11 (수) 벧전4 / 살후1-3
+12.12 (목) 벧전5 / 딤전1-3
+12.13 (금) 벧후1 / 딤전4-6
+12.14 (토) 벧후2 / 딤후1-4
+12.15 (일) 없음 / 딛1-몬1
